Process AOSP wipe commands prior to encryption check
If a password is set on boot in Android, TWRP will stop at the decryption
prompt. If a factory reset or remote wipe command has been issued, this
allows an intruder or thief an opportunity to remove the scripts that
will process these commands. There is also adb access during this prompt,
which means the data block device can be accessed.

By looking for and processing AOSP commands prior to the decryption prompt,
remote wipes will process correctly, and personal data can remain safe.

Change-Id: If0f213398f93f0458b5837ce3e003da5c80e3dc4

Signed-off-by: CheckYourScreen <nimitmehta95@gmail.com>